Refactor area-of-effect modules
