+++ battle_map_tv/area_of_effect/cone.py
@@ -1,86 +1,56 @@
 import math
 from collections import defaultdict
-from typing import List, Tuple, Set
+from typing import Optional, List, Tuple, Set
 
 import numpy as np
+from PySide6.QtCore import QPointF
+from PySide6.QtGui import QPolygonF
+from PySide6.QtWidgets import QGraphicsScene, QGraphicsPolygonItem
 
+from .base_shape import BaseShape
 from battle_map_tv.grid import Grid
 
 
-def circle_to_polygon(
-    x_center: int, y_center: int, radius: int, grid: Grid
-) -> List[Tuple[int, int]]:
-    delta = grid.pixels_per_square
-    radius = radius - radius % delta
-    if radius < delta:
-        return []
-    elif radius < 2 * delta:
-        return [
-            (x_center + radius, y_center + radius),
-            (x_center + radius, y_center - radius),
-            (x_center - radius, y_center - radius),
-            (x_center - radius, y_center + radius),
-        ]
+class Cone(BaseShape):
+    def __init__(
+        self,
+        x1: int,
+        y1: int,
+        x2: int,
+        y2: int,
+        grid: Grid,
+        scene: QGraphicsScene,
+        size: Optional[float] = None,
+    ):
+        self.size = size or self._calculate_size(x1=x1, y1=y1, x2=x2, y2=y2, grid=grid)
+        angle = self._get_angle_radians(x1=x1, y1=y1, x2=x2, y2=y2, grid=grid)
+        point_1, point_2 = calculate_cone_points(point_0=(x1, y1), size=self.size, angle=angle)
+        triangle = QPolygonF.fromList([QPointF(*p) for p in [(x1, y1), point_1, point_2]])
+        self.shape = QGraphicsPolygonItem(triangle)
+        super().__init__(scene=scene)
 
-    edges = CircleEdges()
-    start_point = (0, 0 + radius)
-    edges.add_point(*start_point)
-    x_prev, y_prev = start_point
 
-    while True:
-        x = x_prev + delta
-        y_star = y_prev - delta
-
-        y_circle_prev = math.sqrt(radius**2 - x_prev**2)
-        y_circle = math.sqrt(radius**2 - x**2)
-
-        surface = (x - x_prev) * (y_circle - y_star) + 0.5 * (x - x_prev) * (
-            y_circle_prev - y_circle
+class ConeRasterized(BaseShape):
+    def __init__(
+        self,
+        x1: int,
+        y1: int,
+        x2: int,
+        y2: int,
+        grid: Grid,
+        scene: QGraphicsScene,
+        size: Optional[float] = None,
+    ):
+        self.size: int = int(size or self._calculate_size(x1=x1, y1=y1, x2=x2, y2=y2, grid=grid))
+        angle = self._get_angle_radians(x1=x1, y1=y1, x2=x2, y2=y2, grid=grid)
+        polygon = QPolygonF.fromList(
+            [
+                QPointF(*point)
+                for point in rasterize_cone(x1=x1, y1=y1, size=self.size, angle=angle, grid=grid)
+            ]
         )
-
-        if surface < 0.5 * delta**2:
-            edges.add_point(x_prev, y_star)
-            y = y_star
-        else:
-            y = y_prev
-
-        if x > y:
-            break
-        edges.add_point(x, y)
-        x_prev = x
-        y_prev = y
-
-    points = [(x + x_center, y + y_center) for x, y in edges.get_circle_line()]
-    return points
-
-
-class CircleEdges:
-    def __init__(self):
-        self._edges: List[List[Tuple[int, int]]] = [[] for _ in range(8)]
-
-    def add_point(self, x: int, y: int):
-        points_for_all_octants = [
-            (x, y),
-            (y, x),
-            (y, -x),
-            (x, -y),
-            (-x, -y),
-            (-y, -x),
-            (-y, x),
-            (-x, y),
-        ]
-        for i, point in enumerate(points_for_all_octants):
-            self._edges[i].append(point)
-
-    def get_circle_line(self) -> List[Tuple[int, int]]:
-        final_points = []
-        flip = False
-        for edge in self._edges:
-            if flip:
-                edge = edge[::-1][1:]
-            final_points.extend(edge)
-            flip = False if flip else True
-        return final_points
+        self.shape = QGraphicsPolygonItem(polygon)
+        super().__init__(scene=scene)
 
 
 def rasterize_cone(x1: int, y1: int, size: int, angle: float, grid: Grid) -> List[Tuple[int, int]]:
@@ -210,56 +180,3 @@ def cone_line_segments_to_polygon(
             break
 
     return polygon
-
-
-def round_to_delta(value: float, delta: int) -> int:
-    return delta * round(value / delta)
-
-
-def main():
-    from collections import namedtuple
-    import matplotlib.pyplot as plt
-    import numpy as np
-
-    fig, ax = plt.subplots()
-    size = 30
-    angle = math.radians(-45)
-    delta = 5
-    point_0 = (0, 0)
-
-    point_1, point_2 = calculate_cone_points(point_0=point_0, size=size, angle=angle)
-
-    ax.plot([point_0[0], point_1[0]], [point_0[1], point_1[1]], "r-")
-    ax.plot([point_0[0], point_2[0]], [point_0[1], point_2[1]], "b-")
-    ax.plot([point_1[0], point_2[0]], [point_1[1], point_2[1]], "g--")
-
-    grid = namedtuple("grid", "offset")
-    grid.offset = (0, 0)
-    x_points, y_points = rasterize_cone_by_pixels([point_0, point_1, point_2], delta, grid)  # type: ignore
-    ax.scatter(x_points, y_points, linewidth=3)
-
-    line_segments = cone_pixels_to_line_segments(x_points, y_points, delta)
-    polygon = cone_line_segments_to_polygon(line_segments)
-    ax.plot(*zip(*polygon), "y-")
-
-    ax.xaxis.set_ticks(
-        np.arange(
-            round_to_delta(min([0, point_1[0], point_2[0]]), delta),
-            round_to_delta(max([0, point_1[0], point_2[0]]), delta) + delta,
-            delta,
-        )
-    )
-    ax.yaxis.set_ticks(
-        np.arange(
-            round_to_delta(min([0, point_1[1], point_2[1]]), delta),
-            round_to_delta(max([0, point_1[1], point_2[1]]), delta) + delta,
-            delta,
-        )
-    )
-    ax.grid(True, alpha=0.3)
-    plt.axis("equal")
-    plt.show()
-
-
-if __name__ == "__main__":
-    main()
