+++ battle_map_tv/window_image.py
@@ -4,7 +4,7 @@ from PySide6.QtCore import Qt
 from PySide6.QtGui import QMouseEvent, QImageReader
 from PySide6.QtWidgets import QGraphicsView, QGraphicsScene
 
-from battle_map_tv.aoe import AreaOfEffectManager
+from battle_map_tv.area_of_effect.manager import AreaOfEffectManager
 from battle_map_tv.grid import GridOverlay, Grid
 from battle_map_tv.image import Image
 from battle_map_tv.initiative import InitiativeOverlayManager
